remove old button
