+++ src/index.ts
@@ -4,5 +4,5 @@ export * from './Typography';
 export { default as Grid } from './Grid';
 export * from './Grid';
 
-export { default as Button } from './Button2';
-export * from './Button2';
+export { default as Button } from './Button';
+export * from './Button';
